Refactor spotifyApiCalls util
